Change: Each user has rights to view or change his own Todo only

+++ rest-api/models/todomodel/todomodel.go
@@ -18,6 +18,7 @@ type Todo struct {
 	Name        string `json:"name"`
 	Description string `json:"description"`
 	State       string `json:"state"`
+	UserID      uint64 `json:"user_id" db:"user_id"`
 	CreatedAt   string `json:"created_at" db:"created_at"`
 	UpdatedAt   string `json:"updated_at" db:"updated_at"`
 }
@@ -28,35 +29,35 @@ func NewTodo(db *sqlx.DB) Todos {
 }
 
 // Index method to get all the records in a table
-func (t Todos) Index() ([]Todo, error) {
+func (t Todos) Index(userID string) ([]Todo, error) {
 	todos := []Todo{}
-	sqlStatement := `SELECT todos.id, todos.name, todos.description, states.state, todos.created_at, todos.updated_at 
-	FROM todos LEFT JOIN states ON todos.state=states.id ORDER BY id`
-	err := t.DB.Select(&todos, sqlStatement)
+	sqlStatement := `SELECT todos.id, todos.name, todos.description, todos.user_id, states.state, todos.created_at, todos.updated_at 
+	FROM todos LEFT JOIN states ON todos.state=states.id WHERE todos.user_id=$1 ORDER BY todos.id`
+	err := t.DB.Select(&todos, sqlStatement, userID)
 	return todos, err
 }
 
 // Show method to get a specific record from a table
-func (t Todos) Show(id string) (Todo, error) {
+func (t Todos) Show(todoID string, userID string) (Todo, error) {
 	todo := Todo{}
-	sqlStatement := `SELECT todos.id, todos.name, todos.description, states.state, todos.created_at, todos.updated_at
-	FROM todos LEFT JOIN states ON todos.state=states.id WHERE todos.id=$1`
-	err := t.DB.Get(&todo, sqlStatement, id)
+	sqlStatement := `SELECT todos.id, todos.name, todos.description, todos.user_id, states.state, todos.created_at, todos.updated_at
+	FROM todos LEFT JOIN states ON todos.state=states.id WHERE todos.id=$1 AND todos.user_id=$2`
+	err := t.DB.Get(&todo, sqlStatement, todoID, userID)
 	return todo, err
 }
 
 // Delete method to delete a specific record from a table
-func (t Todos) Delete(id string) (Todo, error) {
+func (t Todos) Delete(todoID string, userID string) (Todo, error) {
 	todo := Todo{}
-	sqlStatement := "DELETE FROM todos WHERE id=$1 RETURNING id, state, name, description, created_at, updated_at"
-	err := t.DB.Get(&todo, sqlStatement, id)
+	sqlStatement := "DELETE FROM todos WHERE id=$1 AND user_id=$2 RETURNING id, state, name, description, user_id, created_at, updated_at"
+	err := t.DB.Get(&todo, sqlStatement, todoID, userID)
 	return todo, err
 }
 
 // Create method to create a record in the table
 func (t Todos) Create(todo Todo) (Todo, error) {
-	sqlStatement := "INSERT INTO todos (name, description, state) VALUES($1, $2, $3) RETURNING id, created_at, updated_at"
-	err := t.DB.Get(&todo, sqlStatement, todo.Name, todo.Description, todo.State)
+	sqlStatement := "INSERT INTO todos (name, description, state, user_id) VALUES($1, $2, $3, $4) RETURNING id, created_at, updated_at"
+	err := t.DB.Get(&todo, sqlStatement, todo.Name, todo.Description, todo.State, todo.UserID)
 	return todo, err
 }
 
@@ -72,7 +73,7 @@ func (t Todos) Update(todo Todo) (Todo, error) {
 	if todo.State != "" {
 		sqlStatement += "state = " + todo.State + ", "
 	}
-	sqlStatement += "updated_at = now() WHERE id=" + strconv.FormatUint(todo.ID, 10) + " RETURNING name, description, state, created_at, updated_at"
+	sqlStatement += "updated_at = now() WHERE id=" + strconv.FormatUint(todo.ID, 10) + " AND user_id=" + strconv.FormatUint(todo.UserID, 10) + " RETURNING name, description, state, created_at, updated_at"
 	err := t.DB.Get(&todo, sqlStatement)
 	return todo, err
 }
